\name{marm3.sim.fbs}
\alias{marm3.sim.fbs}
\docType{package}
\title{
  Generate scenario I data from MARM model.
}
\description{
  Generate scenario I data for MARM model.
}

\usage{
marm3.sim.fbs <- function(n,q,p,s,group=NULL,r10=2,r20=2,r30=2,isfixedR=0,
                           D3=NULL,K=6,degr=3,sigma2=NULL,seed_id=NULL,
                           r1_index=NULL,r2_index=NULL,r3_index=NULL,D0=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{Sample size.}
  
  \item{q}{The number of responses, \eqn{q\geq1}.}
  
  \item{p}{The number of covariates, \eqn{p\geq1}.}
  
  \item{s}{The true covariates associated with responses, \eqn{s\geq1}.}
  
  \item{group}{A length-\eqn{p} vector of the grouping index of predictors, e.g., \eqn{group=c(1,1,1,2,2,2)} means there are \eqn{6} perdictors in the model, and the first three perdictors are in the same group and the last three perdictors are in another one. By default, we set \eqn{group=rep(1,p)}.}
  
  \item{r10}{The first dimension of the tensor. Default is \code{2}.}
  
  \item{r20}{The second dimension of the tensor. Default is \code{2}.}
  
  \item{r30}{The third dimension of the tensor. Default is \code{2}.}
  
  \item{isfixedR}{A logical value indicating whether ranks are fixed.}
  
  \item{D3}{The mode of unfolding \eqn{D_{(3)}}. By default, \code{D3} is generated by random.}
  
  \item{K}{The number of B-spline basis functions, that is the plus of both degrees of basis functioins and the number of knots. Default is \code{6}, which means cubic splines.}
  
  \item{degr}{The number of knots of B-spline base function. Default is \code{3}.}
  
  \item{sigma2}{err variance. Default is \code{0.1}.}
  
  \item{seed_id}{A positive integer, the seed for generating the random numbers. Default is \code{1000}.}
  
  \item{r1_index}{A user-specified sequence of \eqn{r_1} values, where \eqn{r_1} is the first dimension of single value matrix of the tensor. Default is \code{r1_index}\eqn{=1,\cdots,\min(\lceil\log(n)\rceil,p)}. if \eqn{isfixedR=1}, \code{r1_index} is useless.}
            
  \item{r2_index}{A user-specified sequence of \eqn{r_2} values, where \eqn{r_2} is the second dimension of single value matrix of the tensor. Default is \code{r2_index}\eqn{=1,\cdots},\code{max{K_index}}. if \eqn{isfixedR=1}, \code{r2_index} is useless.}
                  
  \item{r3_index}{A user-specified sequence of \eqn{r_3} values, where \eqn{r_3} is the third dimension of single value matrix of the tensor. Default is \code{r3_index}\eqn{=1,\cdots,\min(\lceil\log(n)\rceil,q)}. if \eqn{isfixedR=1}, \code{r3_index} is useless.}
  
  \item{D0}{A user-specified list of initialized values, including \code{ng} sub-lists where ng is the number of groups. For each sub-list, it has four initialized matrices \eqn{S_{(3)}} (called \code{S}), \code{A}, \code{B} and \code{C}. By default, a list of initialization satisfying fixed ranks is computed by random.}
  
}

\details{
This function can generate scenario I data of marm model. \code{isfixedR} is required to yield a different initialization \code{D0}. In scenario I, the true functions are exactly residing in the space of B-spline basis functions.
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
  \item{Y}{Response, a \eqn{n\times q}-matrix.}
  
  \item{X}{Design matrix, a \eqn{n\times p}-matrix.}
  
  \item{f0}{True functions, a \eqn{n\times p}-matrix.}
  
  \item{group}{The grouping index of predictors, a length-\eqn{p} vector.}
  
  \item{D0}{The initialized values.}
  
  \item{...}{Other options for algorithm.}
}

%\author{
%Your Name, email optional.
%Maintainer: Xu Liu <liu.xu@sufe.edu.cn>
%}
\references{
  Multivariate additive regression for multi-view data via tensor estimation.
}
\keyword{ Group sparsity; Tensor low-rankness; Tucker decomposition; Multivariate additive regression; Multi-view data. }
\seealso{
  marm3.sim.fsin
}
\examples{
library(tensorMARM)
n <- 200; q <- 5; p <- 100; s <- 3; ng = 4
group <- rep(1:ng,each=p/ng)
mydata <- marm3.sim.fbs(n,q,p,s,group)

}
