\name{marm4.sim.fsin}
\alias{marm4.sim.fsin}
\docType{package}
\title{
  Generate scenario II data from structural MARM model.
}
\description{
  Generate scenario II data for a structural multivariate additive model for multi-view data.
}

\usage{
marm4.sim.fsin <- function(n,q,p,s,group=NULL,r10=2,r20=2,r30=2,r40=2,
                            isfixedR=0,D42=NULL,K=6,degr=3,sigma2=NULL,
                            seed_id=NULL,r1_index=NULL,r2_index=NULL,
                            r3_index=NULL,r4_index=NULL,D0=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{Sample size.}
  
  \item{q}{The number of responses, \eqn{q\geq1}.}
  
  \item{p}{The number of covariates, \eqn{p\geq1}.}
  
  \item{s}{The true covariates associating to responses, \eqn{s\geq1}.}
  
  \item{group}{A length-\eqn{p} vector of the grouping index of predictors, e.g., \eqn{group=c(1,1,1,2,2,2)} means there are \eqn{6} perdictors in the model, and the first three perdictors are in the same group and the last three perdictors are in another one. By default, we set \eqn{group=c(rep(1,as.integer(p/2)),rep(2,p-as.integer(p/2))}.}
  
  \item{r10}{The first dimension of the tensor. Default is \code{2}.}
  
  \item{r20}{The second dimension of the tensor. Default is \code{2}.}
  
  \item{r30}{The third dimension of the tensor. Default is \code{2}.}
  
  \item{r40}{The third dimension of the tensor. Default is \code{2}.}
  
  \item{isfixedR}{A logical value indicating whether ranks are fixed.}
  
  \item{D42}{The mode of unfolding \eqn{D_{(2)}}. By default, \code{D2} is generated by random.}
  
  \item{K}{The number of B-spline basis functions, that is the plus of both degrees of basis functioins and the number of knots. Default is \code{6}, which means cubic splines.}
  
  \item{degr}{The number of knots of B-spline base function. Default is \code{3}.}
  
  \item{sigma2}{err variance. Default is \code{0.1}.}
  
  \item{seed_id}{A positive integer, the seed for generating the random numbers. Default is \code{1000}.}
  
  \item{r1_index}{A user-specified sequence of \eqn{r_1} values, where \eqn{r_1} is the first dimension of single value matrix of the tensor. Default is \code{r1_index}\eqn{=1,\cdots,\min(\lceil\log(n)\rceil,p)}. if \eqn{isfixedR=1}, \code{r1_index} is useless.}
            
  \item{r2_index}{A user-specified sequence of \eqn{r_2} values, where \eqn{r_2} is the second dimension of single value matrix of the tensor. Default is \code{r2_index}\eqn{=1,\cdots},\code{max{K_index}}. if \eqn{isfixedR=1}, \code{r2_index} is useless.}
                  
  \item{r3_index}{A user-specified sequence of \eqn{r_3} values, where \eqn{r_3} is the third dimension of single value matrix of the tensor. Default is \code{r3_index}\eqn{=1,\cdots,\min(\lceil\log(n)\rceil,ng)}. if \eqn{isfixedR=1}, \code{r3_index} is useless.}
  
  \item{r4_index}{A user-specified sequence of \eqn{r_4} values, where \eqn{r_4} is the third dimension of single value matrix of the tensor. Default is \code{r4_index}\eqn{=1,\cdots,\min(\lceil\log(n)\rceil,q)}. if \eqn{isfixedR=1}, \code{r4_index} is useless.}
  
  \item{D0}{A user-specified list of initialized values, including five initialized matrices \eqn{S_{(4)}} (called \code{S}), \code{A}, \code{B}, \code{C} and \code{D}. By default, a list of initialization satisfying fixed ranks is computed by random.}
  
}

\details{
This function can generate scenario II data of structural marm model. \code{isfixedR} is required to yield a different initialization \code{D0}. In scenario II, the true functions are the linear combination of \eqn{sin(2\pi x)} and \eqn{cos(\pi x)}.
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
  \item{Y}{Response, a \eqn{n\times q}-matrix.}
  
  \item{X}{Design matrix, a \eqn{n\times p}-matrix.}
  
  \item{f0}{True functions, a \eqn{n\times p}-matrix.}
  
  \item{group}{The grouping index of predictors, a length-\eqn{p} vector.}
  
  \item{D0}{The initialized values.}
  
  \item{...}{Other options for algorithm.}
}

%\author{
%Your Name, email optional.
%Maintainer: Xu Liu <liu.xu@sufe.edu.cn>
%}
\references{
  Multivariate additive regression for multi-view data via tensor estimation.
}
\keyword{ Group sparsity; Tensor low-rankness; Tucker decomposition; Multivariate additive regression; Multi-view data. }
\seealso{
  marm4.sim.fbs
}
\examples{
library(tensorMARM)
n <- 200; q <- 5; p <- 100; s <- 3; ng = 4
group <- rep(1:ng,each=p/ng)
mydata <- marm4.sim.fsin(n,q,p,s,group)

}
